pyproject.toml
[tool.poetry]
name = "project-api-department"
version = "0.1.0"
description = ""
authors = ["Viktor Kislyak <viktorkislyak1996@gmail.com>"]
readme = "README.md"
packages = [{include = "project_api_department"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
pre-commit = "^3.3.2"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
